KPieceTable::Piece: handle storageId as enum type across the code

+++ core/piecetable/tests/piecetest.cpp
@@ -34,7 +34,7 @@ void PieceTest::testFullConstructor()
 
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), storageSection.end());
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(piece.isValid());
 }
 
@@ -77,17 +77,17 @@ void PieceTest::testSplitAt()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), Start);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // split at one after start
     piece.set(Start, End);
     splitPiece = piece.splitAt(Start + 1);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), Start);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QCOMPARE(splitPiece.start(), Start + 1);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // split at mid
     const Address Mid = (Start + End) / 2;
@@ -95,27 +95,27 @@ void PieceTest::testSplitAt()
     splitPiece = piece.splitAt(Mid);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), Mid - 1);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QCOMPARE(splitPiece.start(), Mid);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // split at one before width
     piece.set(Start, End);
     splitPiece = piece.splitAt(End);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), End - 1);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QCOMPARE(splitPiece.start(), End);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // split at start so the split is the full
     piece.set(Start, End);
     splitPiece = piece.splitAt(End + 1);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), End);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 }
 
@@ -129,17 +129,17 @@ void PieceTest::testSplitAtLocal()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), Start);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // split at one after start
     piece.set(Start, End);
     splitPiece = piece.splitAtLocal(1);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), Start);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QCOMPARE(splitPiece.start(), Start + 1);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // split at mid
     const Address Mid = Width / 2;
@@ -147,27 +147,27 @@ void PieceTest::testSplitAtLocal()
     splitPiece = piece.splitAtLocal(Mid);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), Start + Mid - 1);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QCOMPARE(splitPiece.start(), Start + Mid);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // split at one before width
     piece.set(Start, End);
     splitPiece = piece.splitAtLocal(Width - 1);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), End - 1);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QCOMPARE(splitPiece.start(), End);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // split at start so the split is the full
     piece.set(Start, End);
     splitPiece = piece.splitAtLocal(Width);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), End);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 }
 
@@ -183,7 +183,7 @@ void PieceTest::testRemove()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), Start);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove one at start
     piece.set(Start, End);
@@ -192,7 +192,7 @@ void PieceTest::testRemove()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), Start + 1);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove many at start
     piece.set(Start, End);
@@ -201,7 +201,7 @@ void PieceTest::testRemove()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), Mid + 1);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove all except last
     piece.set(Start, End);
@@ -210,7 +210,7 @@ void PieceTest::testRemove()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), End);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove at mid
     piece.set(Start, End);
@@ -218,10 +218,10 @@ void PieceTest::testRemove()
     splitPiece = piece.remove(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), Mid - 2);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QCOMPARE(splitPiece.start(), Mid + 2);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove none at width
     piece.set(Start, End);
@@ -229,7 +229,7 @@ void PieceTest::testRemove()
     splitPiece = piece.remove(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(),   End);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 
     // remove one at width
@@ -238,7 +238,7 @@ void PieceTest::testRemove()
     splitPiece = piece.remove(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(),   End - 1);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 
     // remove many at width
@@ -247,7 +247,7 @@ void PieceTest::testRemove()
     splitPiece = piece.remove(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(),   Mid - 1);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 
     // remove all except first
@@ -256,7 +256,7 @@ void PieceTest::testRemove()
     splitPiece = piece.remove(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(),   Start);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 
     // remove all
@@ -279,7 +279,7 @@ void PieceTest::testRemoveLocal()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), Start);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove one at start
     piece.set(Start, End);
@@ -288,7 +288,7 @@ void PieceTest::testRemoveLocal()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), Start + 1);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove many at start
     piece.set(Start, End);
@@ -297,7 +297,7 @@ void PieceTest::testRemoveLocal()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), Start + Mid + 1);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove all except last
     piece.set(Start, End);
@@ -306,7 +306,7 @@ void PieceTest::testRemoveLocal()
     QVERIFY(!piece.isValid());
     QCOMPARE(splitPiece.start(), End);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove at mid
     piece.set(Start, End);
@@ -314,10 +314,10 @@ void PieceTest::testRemoveLocal()
     splitPiece = piece.removeLocal(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(), Start + Mid - 2);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QCOMPARE(splitPiece.start(), Start + Mid + 2);
     QCOMPARE(splitPiece.end(),   End);
-    QCOMPARE(splitPiece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(splitPiece.storageId(), Piece::ChangeStorage);
 
     // remove none at width
     piece.set(Start, End);
@@ -325,7 +325,7 @@ void PieceTest::testRemoveLocal()
     splitPiece = piece.removeLocal(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(),   End);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 
     // remove one at width
@@ -334,7 +334,7 @@ void PieceTest::testRemoveLocal()
     splitPiece = piece.removeLocal(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(),   End - 1);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 
     // remove many at width
@@ -343,7 +343,7 @@ void PieceTest::testRemoveLocal()
     splitPiece = piece.removeLocal(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(),   Start + Mid - 1);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 
     // remove all except first
@@ -352,7 +352,7 @@ void PieceTest::testRemoveLocal()
     splitPiece = piece.removeLocal(RemoveSection);
     QCOMPARE(piece.start(), Start);
     QCOMPARE(piece.end(),   Start);
-    QCOMPARE(piece.storageId(), (int)Piece::ChangeStorage);
+    QCOMPARE(piece.storageId(), Piece::ChangeStorage);
     QVERIFY(!splitPiece.isValid());
 
     // remove all
